Agregué la rutina de interrución del P1.3, ademas de seleccionar el disparo por flanco descendiente. Activadas en el nRF las interuciones por TX_DS y MAX_RT en la fucnion nRF init. Terminada la función enviar dato.

+++ CCS/TX/variables.h
@@ -10,7 +10,7 @@
 
 //generales
 
-uint16_t i	= 0;
+uint8_t i	= 0;
 uint16_t dato;
